Configurar Trust Proxies para Railway

+++ database/migrations/2025_11_06_093701_create_coleccion_imagen_table.php
@@ -7,16 +7,13 @@ return new class extends Migration
 {
     public function up(): void
     {
-        // Recomendación: Usa minúsculas para el nombre de la tabla también: 'coleccion_imagen'
         if (!Schema::hasTable('coleccion_imagen')) {
             Schema::create('coleccion_imagen', function (Blueprint $table) {
-                $table->unsignedInteger('id_coleccion');
+                // CORRECCIÓN: Cambiado a unsignedBigInteger para coincidir con el padre
+                $table->unsignedBigInteger('id_coleccion');
                 $table->unsignedBigInteger('id_imagen');
 
-                // CORRECCIÓN: Referencias en minúscula
                 $table->foreign('id_coleccion')->references('id_coleccion')->on('coleccion')->onDelete('cascade');
-
-                // CORRECCIÓN: Referencias en minúscula
                 $table->foreign('id_imagen')->references('id_imagen')->on('imagen')->onDelete('cascade');
 
                 $table->primary(['id_coleccion', 'id_imagen']);
